+++ index.html
@@ -12,7 +12,7 @@
 
 <body>
   <div id="map"></div>
-  <script src="dist/main.js"></script>
+  <script src="src/index.js"></script>
 </body>
 
 </html>
